0.6.6
-Stacks of items in the deconstruct window will now always be YELLOW. If
it is the last trait of its kind when it reaches a stack size of 1, it
will turn orange.
-Profession tracking options are now character specific and have earned
their own place in the settings menu. I changed this in such a way that
it won't reset the rest of your settings, but the three tracking
settings will be changed to ON upon applying this update.
-Added french translations thanks to Ykses